Fix problem in state estimation
It seems that there is a bug in computing the S matrix that affects the
compulation of the derivative matrix H. As, such a new function has
been developed to compute S. Moreover, the code is hardwired to solve
the problem 6.7 in the book "Computational Methods for Electric Power
Systems" by Mariesa Crow, 1st edition. In particular, it only accepts
the measurements in that problem. So, the code is expanded to
accommodate any sets of measurements. A test file "test_se_Ex_6_17.m"
is provided that solves example 6.17 in the above 2nd edit of the book.
The file shows the difference in computing H between the original,
which is edited to allow more measurements, and the modified code.
The modified code produced results that coincide with the solution of
example 6.17 in the book.